Refactor RPN in a separate function.
