+++ keras_rcnn/models/rcnn/__init__.py
@@ -1,0 +1,1 @@
+from .resnet import ResNet50RCNN
